+++ guidata/guitest.py
@@ -25,17 +25,20 @@ from qtpy.QtWidgets import (
     QGroupBox,
     QHBoxLayout,
     QShortcut,
+    QMainWindow,
+    QFrame,
 )
 from qtpy.QtGui import QKeySequence
 from qtpy.QtCore import Qt, QSize
 
 from guidata.config import _
 from guidata.configtools import get_icon
-from guidata.qthelpers import get_std_icon
+from guidata.qthelpers import get_std_icon, win32_fix_title_bar_background
 from guidata.widgets.codeeditor import PythonCodeEditor
 
 
 def get_tests(test_package):
+    """Retrieve test scripts from test package"""
     tests = []
     test_path = osp.dirname(osp.realpath(test_package.__file__))
     for fname in sorted(os.listdir(test_path)):
@@ -52,6 +55,8 @@ def get_tests(test_package):
 
 
 class TestModule(object):
+    """Object representing a test module (Python script)"""
+
     def __init__(self, test_module):
         self.module = test_module
         self.filename = osp.splitext(osp.abspath(test_module.__file__))[0] + ".py"
@@ -59,19 +64,22 @@ class TestModule(object):
             self.filename += "w"
 
     def is_visible(self):
+        """Returns True if this script is intended to be shown in test launcher"""
         return hasattr(self.module, "SHOW") and self.module.SHOW
 
     def get_description(self):
+        """Returns test module description"""
         doc = self.module.__doc__
         if doc is None or not doc.strip():
             return _("No description available")
         else:
             lines = doc.strip().splitlines()
-            format = "<span style='color: #2222FF'><b>%s</b></span>"
-            lines[0] = format % lines[0]
+            fmt = "<span style='color: #2222FF'><b>%s</b></span>"
+            lines[0] = fmt % lines[0]
             return "<br>".join(lines)
 
     def run(self, args=""):
+        """Run test script"""
         # Keep the same sys.path environment in child process:
         # (useful when the program is executed from Spyder, for example)
         os.environ["PYTHONPATH"] = os.pathsep.join(sys.path)
@@ -83,90 +91,93 @@ class TestModule(object):
 
 
 class TestPropertiesWidget(QWidget):
+    """Test module properties panel"""
+
     def __init__(self, parent):
         QWidget.__init__(self, parent)
         info_icon = QLabel()
         icon = get_std_icon("MessageBoxInformation").pixmap(24, 24)
         info_icon.setPixmap(icon)
         info_icon.setFixedWidth(32)
-        info_icon.setAlignment(Qt.AlignTop)
         self.desc_label = QLabel()
         self.desc_label.setWordWrap(True)
-        self.desc_label.setAlignment(Qt.AlignTop)
         group_desc = QGroupBox(_("Description"), self)
         layout = QHBoxLayout()
-        layout.addWidget(info_icon)
-        layout.addWidget(self.desc_label)
+        for label in (info_icon, self.desc_label):
+            label.setAlignment(Qt.AlignTop)
+            layout.addWidget(label)
         group_desc.setLayout(layout)
 
-        self.editor = PythonCodeEditor(self)
-        self.desc_label.setFont(self.editor.font())
+        self.editor = PythonCodeEditor(self, columns=80, rows=30)
         self.editor.setReadOnly(True)
-        group_code = QGroupBox(_("Source code"), self)
-        layout = QVBoxLayout()
-        layout.addWidget(self.editor)
-        group_code.setLayout(layout)
-
-        self.run_button = QPushButton(get_icon("apply.png"), _("Run this script"), self)
-        self.quit_button = QPushButton(get_icon("exit.png"), _("Quit"), self)
-        hlayout = QHBoxLayout()
-        hlayout.addWidget(self.run_button)
-        hlayout.addStretch()
-        hlayout.addWidget(self.quit_button)
+        self.desc_label.setFont(self.editor.font())
 
         vlayout = QVBoxLayout()
         vlayout.addWidget(group_desc)
-        vlayout.addWidget(group_code)
-        vlayout.addLayout(hlayout)
+        vlayout.addWidget(self.editor)
         self.setLayout(vlayout)
 
     def set_item(self, test):
+        """Set current item"""
         self.desc_label.setText(test.get_description())
         self.editor.set_text_from_file(test.filename)
 
 
-class TestLauncherWindow(QSplitter):
+class TestMainView(QSplitter):
+    """Test launcher main view"""
+
     def __init__(self, package, parent=None):
         QSplitter.__init__(self, parent)
-        self.setWindowTitle(_("Tests - %s module") % package.__name__)
-        self.setWindowIcon(get_icon("%s.svg" % package.__name__, "guidata.svg"))
-
         test_package_name = "%s.tests" % package.__name__
         _temp = __import__(test_package_name)
         test_package = sys.modules[test_package_name]
 
         tests = get_tests(test_package)
-        listwidget = QListWidget(self)
-        listwidget.addItems([osp.basename(test.filename) for test in tests])
+        listgroup = QFrame()
+        props = TestPropertiesWidget(self)
+        font = props.editor.font()
+        self.addWidget(listgroup)
+        self.addWidget(props)
 
-        self.properties = TestPropertiesWidget(self)
+        listw = QListWidget(self)
+        listw.addItems([osp.basename(test.filename) for test in tests])
+        for index in range(listw.count()):
+            listw.item(index).setSizeHint(QSize(1, 25))
+        listw.setFont(font)
+        listw.currentRowChanged.connect(lambda row: props.set_item(tests[row]))
+        listw.itemActivated.connect(lambda: tests[listw.currentRow()].run())
+        listw.setCurrentRow(0)
+        run_button = QPushButton(get_icon("apply.png"), _("Run this script"), self)
+        run_button.setFont(font)
+        run_button.clicked.connect(lambda: tests[listw.currentRow()].run())
 
-        self.addWidget(listwidget)
-        self.addWidget(self.properties)
+        vlayout = QVBoxLayout()
+        vlayout.addWidget(listw)
+        vlayout.addWidget(run_button)
+        listgroup.setLayout(vlayout)
 
-        self.properties.run_button.clicked.connect(
-            lambda: tests[listwidget.currentRow()].run()
-        )
-        self.properties.quit_button.clicked.connect(self.close)
-        listwidget.currentRowChanged.connect(
-            lambda row: self.properties.set_item(tests[row])
-        )
-        listwidget.itemActivated.connect(lambda: tests[listwidget.currentRow()].run())
-        listwidget.setCurrentRow(0)
-
-        QShortcut(QKeySequence("Escape"), self, self.close)
-
-        self.setSizes([150, 1])
         self.setStretchFactor(1, 1)
-        self.resize(QSize(840, 600))
-        self.properties.set_item(tests[0])
+        props.set_item(tests[0])
+
+
+class TestLauncherWindow(QMainWindow):
+    """Test launcher main window"""
+
+    def __init__(self, package, parent=None):
+        QMainWindow.__init__(self, parent)
+        win32_fix_title_bar_background(self)
+        self.setWindowTitle(_("Tests - %s module") % package.__name__)
+        self.setWindowIcon(get_icon("%s.svg" % package.__name__, "guidata.svg"))
+        self.mainview = TestMainView(package, self)
+        self.setCentralWidget(self.mainview)
+        QShortcut(QKeySequence("Escape"), self, self.close)
 
 
 def run_testlauncher(package):
     """Run test launcher"""
-    from qtpy.QtWidgets import QApplication
+    from guidata import qapplication
 
-    app = QApplication([])
+    app = qapplication()
     win = TestLauncherWindow(package)
     win.show()
     app.exec_()